backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "law-rag-backend"
version = "0.1.0"
description = "Law RAG Application Backend - MVP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic-settings>=2.1.0",
    "structlog>=24.1.0",
    "sqlalchemy>=2.0.25",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "python-multipart>=0.0.6",
    "pgvector>=0.2.0",
    "psycopg2-binary>=2.9.9",
    "sentence-transformers>=2.2.0",
    "google-genai>=1.0.0",
    "neo4j>=5.14.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "pyyaml>=6.0",
]

[project.scripts]
law-rag-eval = "app.cli_eval:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
pythonpath = ["."]

[tool.hatch.build.targets.wheel]
packages = ["app"]
